extract channel history mock
