+++ tests/cogs/test_insights.py
@@ -2,6 +2,7 @@ import pytest
 import mock
 import datetime
 from async_mock_ext import patch_async_mock
+from mock_channel_history import ChannelHistory
 import server.channels as channels
 from cogs.insights import Insights
 
@@ -11,7 +12,7 @@ from cogs.insights import Insights
 @mock.patch('discord.TextChannel')
 async def test_check_should_respond_no_messages(bot, channel):
     bot.get_channel.return_value = channel
-    channel.history.return_value = Hist(None)
+    channel.history.return_value = ChannelHistory(None)
     clazz = Insights(bot, start_tasks=False)
     await clazz._Insights__check_should_respond()
     bot.get_channel.assert_called_once_with(channels.GENERAL)
@@ -29,7 +30,7 @@ async def test_check_should_respond_new_message(bot, channel, message):
         bot.get_channel.return_value = channel
         message.created_at = datetime.datetime(2000, 1, 1, hour = 11, minute = 38)
         message.author.id = 244629273191645184
-        channel.history.return_value = Hist(message)
+        channel.history.return_value = ChannelHistory(message)
         clazz = Insights(bot, start_tasks=False)
         await clazz._Insights__check_should_respond()
         bot.get_channel.assert_called_once_with(channels.GENERAL)
@@ -47,7 +48,7 @@ async def test_check_should_respond_not_special_user(bot, channel, message):
         bot.get_channel.return_value = channel
         message.created_at = datetime.datetime(2000, 1, 1, hour = 11, minute = 00)
         message.author.id = 0
-        channel.history.return_value = Hist(message)
+        channel.history.return_value = ChannelHistory(message)
         clazz = Insights(bot, start_tasks=False)
         await clazz._Insights__check_should_respond()
         bot.get_channel.assert_called_once_with(channels.GENERAL)
@@ -65,18 +66,8 @@ async def test_check_should_respond_old_message_sent_by_special_user(bot, channe
         bot.get_channel.return_value = channel
         message.created_at = datetime.datetime(2000, 1, 1, hour = 11, minute = 00)
         message.author.id = 244629273191645184
-        channel.history.return_value = Hist(message)
+        channel.history.return_value = ChannelHistory(message)
         clazz = Insights(bot, start_tasks=False)
         await clazz._Insights__check_should_respond()
         bot.get_channel.assert_called_once_with(channels.GENERAL)
         channel.send.assert_called()
-
-class Hist:
-    def __init__(self, message):
-        self.message = message
-    
-    async def flatten(self):
-        if self.message:
-            return [ self.message ]
-        else:
-            return []
